Moving all area data except the lut into a non fixed bank. This data can get pretty large as this game grows. The bank number for the areas is hard-coded at the moment. It shouldn't be hard to un-hard code it later if we really need to.
Modifying play_state to switch to the areas bank whenever we're reading data from an area, and switch to other data right before we need it. It turned out not to be so difficult to do this.

Adding a variable to ram for textbox_attribute, so we don't have to switch to the area bank just to load this information when showing a textbox.

git-svn-id: svn://localhost/Owlia/trunk@170 c1cc9e3d-c02f-ee42-9970-117c3f37ddad