Modifying sprite_hide_all_below to not just hide but also align sprites with the nametable, by rounding to the closest multiple of 8 for y coordinates. Moving the call to this routine to the play state, right after we have aligned with the nametable and before the textbox is first drawn, to guarantee sprites are immediately hidden by the textbox.
This is an improvement over the current behavior which just leaves ugly gaps when hiding sprites behind the textbox. The trade off is a slight "pop" in the y coordinate of the hero and any sprites that are partially occluded by the textbox. In the future we might be able to improve this a bit further by somehow telling all entities to align themselves to an 8 pixel boundary on a frame by frame basis. I'm not sure which will be better. No matter what we choose here it will not be 100% ideal, but this is pretty good for now, I think.

git-svn-id: svn://localhost/Owlia/trunk@180 c1cc9e3d-c02f-ee42-9970-117c3f37ddad