Adding screen coordinates to the hero, familiar, and non player entities. Adding routines for calculating screen coordinates and adding calls to these routines. Adding routines for aligning entities with the nametable when they intersect the textbox. Removing logic from the sprite_hide_all_below routine that originally aligned sprites to the nametable boundary.
The purpose of this code was to only align entities with the nametable when starting a conversation when they intersect the textbox, while at the same time not introducing much overhead. The goal is to minimize the "pop" of the entities as we try to make them appear to be hidden behind the textbox during a conversation.

git-svn-id: svn://localhost/Owlia/trunk@189 c1cc9e3d-c02f-ee42-9970-117c3f37ddad