Adding scroll glitch hiding using cycle timed vblank code. This works by padding cpu cycles in the map row and column update routines when one or both of them is a no-op. And, we had to pad the palette upload routine for transitions. Then, after ppu updates have been made and a frame has started we turn off rendering, pad more cpu cycles to wait 16 scanlines, and then turn rendering back on. We do this only for background rendering. Doing it for both bg and sprites would introduce what is called the "alternate dot crawl" bug in the NES.
We had to adjust how sprites get clipped, and change the camera origin and textbox position to ensure that all continued to work the same.

We had to change location macros to adjust scrolling properly for the new scheme.

Camera y increment routines were adjusted to reflect default scrolling values near the edges of a map.asm

A bug shook out of this where we had to avoid uploading sprite OAM when filling the nametable when loading a location. It seemed to cause nametable corruption very intermittently, though I do not know for sure why.



git-svn-id: svn://localhost/Owlia/trunk@242 c1cc9e3d-c02f-ee42-9970-117c3f37ddad