These changes are intended to hopefully make bankswitching between the main thread and the nmi thread more robust. I believe it is possible for a race condition to exist where we switch to another bank, interrupt the main thread, load and restore the current bank BEFORE storing it in the main thread, return to the main thread and see that the bank is wrong.
Also, it appears to be bad practice to interrupt the main thread with bankswitching at all---so I'm trying to reduce how much interruptible bankswitching can occur in the main thread while loading a full screen of nametable graphics.

Finally, it also appears that messing with current_bank in nmi and saving and restoring it from the stack may cause some issues, so I'm simplifying safe_soundengine_update as a result.

I'm afraid there is still a problem here. I have some patches stored which can reveal this nametable glitch bug, which happens while music is running and we are loading graphics. I wish I knew what the problem was but I'm really at a loss to truly understand what the problem is. I may have to be happy with empircal testing revealing little or no faults, in this case. By checking this in, I'll probably jinx the bug bag into existence.

git-svn-id: svn://localhost/Owlia/trunk@243 c1cc9e3d-c02f-ee42-9970-117c3f37ddad