Adding new print_string_impl and macro print_string for convenient string printing. Adding ES control character for end of string control.
Adding a constant for the chr bank that the textbox chr data resides in, and adding code to switch to this bank whenever we're loading textbox and font data.

Adding a test string to the inventory state.

Updating inventory screen graphics to have a crappy looking border to suggest the sort of look we're going for when it is drawn to look nice.

git-svn-id: svn://localhost/Owlia/trunk@262 c1cc9e3d-c02f-ee42-9970-117c3f37ddad