Implementing the keyed door entity. This entity takes advantage of the new floating solid metatile functionality. It also short circuits the hero's A button attack the same way the NPC does by calling hero_cancel_attack. It only lets you open the door if you are walking towards it and hitting A at the same time. This works because the hero's rectangle will be offset slightly in this case, until it is ejected AFTER entities are finished running their logic. We had to modify the order in which buttons are tested in the hero logic because previously we were skipping the direction handlers whenever the A button was detected, breaking the desired operation of the keyed doors. It worked by coincidence without this change in a couple of directions because rectangles are too large on the right and on the bottom, so northern and western doors worked without the change. I opened issue OWLIA-25 for fixing that.
I also changed the makefile to build debug files for the new code segment which starts at $C400. I should probably add another for the vblank segment, too. I modified the hitbox lua script to help me debug the keyed doors but for some reason I don't get a rectangle for every call to rect_in_rect. Not sure why. Various bits of data in the dungeon area were modified for testing. I removed the ability for locations to be locked, and removed the logic for unlocking locations.

I modified entity_grab_floating_solid_metatile to take parameters instead of always using the entity's coordinates. Then I was able to parameterize keyeddoor entities to grab the tile to the top half or the bottom half, making it possible to create southern doors.

I also removed the check for the floating metatile being at 0 deciding it was best to simply ensure that it is out of the way whenever an entity no longer needs it. I did this because I ended up testing a door that had coordinates 0,7, and that check made it impossible to use at that location.

git-svn-id: svn://localhost/Owlia/trunk@292 c1cc9e3d-c02f-ee42-9970-117c3f37ddad