Refactoring the name of vblank_wait_flag to be vblank_done_flag. Changing the meaning of this flag so that 0 means "not done" and 1 means "done." This is closer to my intuitive understanding of how this flag is used. We clear it, and then wait for it to be done when we want to ensure we are within a safe window where we won't get interrupted, either for changes to the PPU or the indirect vblank_routine variable.
Adding a macro, safely_set_vblank_routine, which waits until a vblank has completed to ensure that the vblank_routine address can be changed without being interrupted between bytes (this could easily cause a crash).

I think we're still being sloppy with sprite data and also palette data. We might want to have ready flags for these just as we have ready flags for uploading a row or a column of tiles in the map ppu routine.

git-svn-id: svn://localhost/Owlia/branches/new_dungeon_exits_approach2@699 c1cc9e3d-c02f-ee42-9970-117c3f37ddad