OWLIA-151: Modifying the textbox engine to draw a smaller textbox. It is smaller by 16 pixels on the left and the right side. This should look better particularly on TVs where it can get cut off by a couple of pixels on the sides. The only side effect I anticipate from this is if there would ever be a situation with a sprite that gets cut off in the areas that are not covered by the textbox. Obviously we're not going to bother trying to both horizontally and vertically align sprites just to account for these (hopefully) rare edge cases. If it ever happens we'll consider it a reasonable trade off for getting a more attractive textbox.
In order to properly decode nametable and attribute table data for cut scenes, we had to introduce a blank map. This is because previously, we actually were decoding whatever garbage all the map addresses were pointing to and then overwriting said garbage for an entire row, so it worked out. Now, because we brought the textbox in at the sides, there are gaps. Instead of going the long and hard route and separating out more logic in the map engine I just added a blank map: all 0's. This would only break if the 0'th tile in the loaded CHR ram was not the blank tile. But I don't anticipate that happening for cut scenes.

Finally, we had to modify all current conversations to fit within the smaller textbox, which was quite tedious, but necessary. I sure hope I'll never have to do that again. I don't anticipate that I will; this should be the final way the textbox looks.

git-svn-id: svn://localhost/Owlia/trunk@770 c1cc9e3d-c02f-ee42-9970-117c3f37ddad